修复退出异常，删除异常

+++ src/views/coach/index.vue
@@ -69,6 +69,7 @@
 
 <script>
 import { deleteById, getList, addCoach, updateCoach } from '@/api/coach'
+import { orderExist } from '@/api/order'
 export default {
     data() {
         return {
@@ -94,6 +95,30 @@ export default {
             textMap: {
                 update: 'Edit',
                 create: 'Create'
+            },
+            rules: {
+                memberId: [
+                    { required: true, message: '请输入会员id', trigger: 'blur' }
+                ],
+                coachId: [
+                    { required: true, message: '请输入教练id', trigger: 'blur' }
+                ],
+                fieldId: [
+                    { required: true, message: '请输入场地id', trigger: 'blur' }
+                ],
+                courseId: [
+                    { required: true, message: '请输入课程id', trigger: 'blur' }
+                ],
+                startDate: [
+                    { type: 'date', required: true, message: '请选择日期', trigger: 'change' }
+                ],
+                startTime: [
+                    { type: 'time', required: true, message: '请选择时间', trigger: 'change' }
+                ],
+                endTime: [
+                    { type: 'time', required: true, message: '请选择时间', trigger: 'change' }
+                ],
+
             }
         }
     },
@@ -192,15 +217,31 @@ export default {
             })
         },
         handleDelete(row){
-            deleteById(row).then(responce=>{
-                this.$notify({
-                    title: '成功',
-                    message: '删除成功',
-                    type: 'success',
-                    duration: 2000
-                })
-                this.fetchData()
+            orderExist({
+                flag:2,
+                id:row.id
+            }).then(responce=>{
+                if(responce.data.sum!=0){
+                    this.$notify({
+                        title: '警告',
+                        message: '存在有效订单',
+                        type: 'warning',
+                        duration: 2000
+                    })
+                }else{
+                    deleteById(row).then(responce => {
+                        this.$notify({
+                            title: '成功',
+                            message: '删除成功',
+                            type: 'success',
+                            duration: 2000
+                        })
+                        this.fetchData()
+                    })
+                }
+
             })
+            
         }
     }
 }
